6. app_Ai/requirements.txt
Flask
pandas
numpy
python-dateutil
timezonefinder
meteostat
pvlib
geocoder
